requirements.txt
matplotlib==3.3.4
numpy==1.21.5
opencv_python==4.5.1.48
Pillow==9.4.0
tensorflow==2.12.0
tensorflow_intel==2.11.0
torch==1.9.0
torchvision==0.9.0
